Test randomness. Move IdentityTransformation
