+++ tests/Refinery/Random/Transformation/ShuffleTransformationTest.php
@@ -3,7 +3,7 @@
 /**
  * @author  Lukas Scharmer <lscharmer@databay.de>
  */
-namespace ILIAS\Tests\Refinery\Random\Transformation;
+namespace ILIAS\Tests\Refinery;
 
 use ILIAS\Refinery\Random\Transformation\ShuffleTransformation;
 use ILIAS\Refinery\Random\Seed\Seed;
@@ -16,12 +16,16 @@ class ShuffleTransformationTest extends TestCase
 {
     public function testTransformResultSuccess() : void
     {
-        $value = ['nrrrrg'];
+        $seed = 0;
+        $value = ['Donec', 'at', 'pede', 'Phasellus', 'purus', 'Nulla', 'facilisis', 'risus', 'a', 'rhoncus', 'fermentum', 'tellus', 'tellus', 'lacinia', 'purus', 'et', 'dictum', 'nunc', 'justo', 'sit', 'amet', 'elit'];
+        $expected = $this->shuffleWithSeed($value, $seed);
         $seedMock = $this->getMockBuilder(Seed::class)->getMock();
-        $seedMock->expects(self::once())->method('seedRandomGenerator');
+        $seedMock->expects(self::once())->method('seedRandomGenerator')->willReturnCallback(static function () use ($seed) : void {
+            \mt_srand($seed);
+        });
 
         $result = (new ShuffleTransformation($seedMock))->transform($value);
-        $this->assertEquals($value, $result);
+        $this->assertEquals($expected, $result);
     }
 
     public function testTransformResultFailure() : void
@@ -31,5 +35,13 @@ class ShuffleTransformationTest extends TestCase
         $seedMock->expects(self::never())->method('seedRandomGenerator');
 
         $result = (new ShuffleTransformation($seedMock))->transform('im no array');
+    }
+
+    private function shuffleWithSeed(array $array, int $seed) : array
+    {
+        \mt_srand($seed);
+        \shuffle($array);
+
+        return $array;
     }
 }
